+++ tests/json-api/tests/integration/cache/data-documents-test.ts
@@ -7,7 +7,6 @@ import Store from '@ember-data/store';
 import type { NotificationType } from '@ember-data/store/-private/managers/notification-manager';
 import type {
   CollectionResourceDataDocument,
-  ResourceMetaDocument,
   SingleResourceDataDocument,
   StructuredDocument,
 } from '@ember-data/types/cache/document';
@@ -70,7 +69,7 @@ class TestSchema<T extends string> {
   }
 }
 
-module('Integration | @ember-data/json-api Cache', function (hooks) {
+module('Integration | @ember-data/json-api Cache.put(<DataDocument>)', function (hooks) {
   setupTest(hooks);
 
   hooks.beforeEach(function () {
@@ -178,200 +177,6 @@ module('Integration | @ember-data/json-api Cache', function (hooks) {
       {
         lid: 'https://api.example.com/v1/users',
         data: [identifier, identifier2],
-      },
-      'We got the cached response document back'
-    );
-  });
-
-  test('meta documents are correctly updated', function (assert) {
-    const store = this.owner.lookup('service:store') as Store;
-
-    const responseDocument = store.cache.put({
-      request: { url: 'https://api.example.com/v1/users' },
-      content: {
-        meta: { count: 4, last: 4 },
-      },
-    } as StructuredDocument<ResourceMetaDocument>) as ResourceMetaDocument;
-
-    assert.false('data' in responseDocument, 'No data is associated');
-    assert.deepEqual(responseDocument.meta, { count: 4, last: 4 }, 'meta is correct');
-    assert.strictEqual(JSON.stringify(responseDocument.meta), JSON.stringify({ count: 4, last: 4 }), 'meta is correct');
-    assert.strictEqual(responseDocument.lid, 'https://api.example.com/v1/users', 'lid is correct');
-
-    const structuredDocument = store.cache.peekRequest({ lid: 'https://api.example.com/v1/users' });
-    assert.deepEqual(
-      structuredDocument,
-      {
-        request: { url: 'https://api.example.com/v1/users' },
-        content: {
-          lid: 'https://api.example.com/v1/users',
-          meta: { count: 4, last: 4 },
-        },
-      },
-      'We got the cached structured document back'
-    );
-    const cachedResponse = store.cache.peek({ lid: 'https://api.example.com/v1/users' });
-    assert.deepEqual(
-      cachedResponse,
-      {
-        lid: 'https://api.example.com/v1/users',
-        meta: { count: 4, last: 4 },
-      },
-      'We got the cached response document back'
-    );
-
-    const responseDocument2 = store.cache.put({
-      request: { url: 'https://api.example.com/v1/users' },
-      content: {
-        meta: { count: 3, next: 8 },
-      },
-    } as StructuredDocument<ResourceMetaDocument>) as ResourceMetaDocument;
-
-    assert.false('data' in responseDocument2, 'No data is associated');
-    assert.deepEqual(responseDocument2.meta, { count: 3, next: 8 }, 'meta is correct');
-    assert.strictEqual(
-      JSON.stringify(responseDocument2.meta),
-      JSON.stringify({ count: 3, next: 8 }),
-      'meta is correct'
-    );
-    assert.strictEqual(responseDocument2.lid, 'https://api.example.com/v1/users', 'lid is correct');
-
-    const structuredDocument2 = store.cache.peekRequest({ lid: 'https://api.example.com/v1/users' });
-    assert.deepEqual(
-      structuredDocument2,
-      {
-        request: { url: 'https://api.example.com/v1/users' },
-        content: {
-          lid: 'https://api.example.com/v1/users',
-          meta: { count: 3, next: 8 },
-        },
-      },
-      'We got the cached structured document back'
-    );
-    const cachedResponse2 = store.cache.peek({ lid: 'https://api.example.com/v1/users' });
-    assert.deepEqual(
-      cachedResponse2,
-      {
-        lid: 'https://api.example.com/v1/users',
-        meta: { count: 3, next: 8 },
-      },
-      'We got the cached response document back'
-    );
-  });
-
-  test('updating cache with a meta document disregards prior data', function (assert) {
-    const store = this.owner.lookup('service:store') as Store;
-
-    const responseDocument = store.cache.put({
-      request: { url: 'https://api.example.com/v1/users' },
-      content: {
-        data: [{ type: 'user', id: '1', attributes: { name: 'Chris' } }],
-        meta: { count: 4, last: 4 },
-      },
-    } as StructuredDocument<CollectionResourceDocument>) as CollectionResourceDataDocument;
-    const identifier = store.identifierCache.getOrCreateRecordIdentifier({ type: 'user', id: '1' });
-
-    assert.deepEqual(responseDocument.data, [identifier], 'data is associated');
-    assert.deepEqual(responseDocument.meta, { count: 4, last: 4 }, 'meta is correct');
-    assert.strictEqual(JSON.stringify(responseDocument.meta), JSON.stringify({ count: 4, last: 4 }), 'meta is correct');
-    assert.strictEqual(responseDocument.lid, 'https://api.example.com/v1/users', 'lid is correct');
-
-    const structuredDocument = store.cache.peekRequest({ lid: 'https://api.example.com/v1/users' });
-    assert.deepEqual(
-      structuredDocument,
-      {
-        request: { url: 'https://api.example.com/v1/users' },
-        content: {
-          lid: 'https://api.example.com/v1/users',
-          data: [identifier],
-          meta: { count: 4, last: 4 },
-        },
-      },
-      'We got the cached structured document back'
-    );
-    const cachedResponse = store.cache.peek({ lid: 'https://api.example.com/v1/users' });
-    assert.deepEqual(
-      cachedResponse,
-      {
-        lid: 'https://api.example.com/v1/users',
-        data: [identifier],
-        meta: { count: 4, last: 4 },
-      },
-      'We got the cached response document back'
-    );
-
-    const responseDocument2 = store.cache.put({
-      request: { url: 'https://api.example.com/v1/users' },
-      content: {
-        meta: { count: 3, next: 8 },
-      },
-    } as StructuredDocument<ResourceMetaDocument>) as ResourceMetaDocument;
-
-    assert.false('data' in responseDocument2, 'No data is associated');
-    assert.deepEqual(responseDocument2.meta, { count: 3, next: 8 }, 'meta is correct');
-    assert.strictEqual(
-      JSON.stringify(responseDocument2.meta),
-      JSON.stringify({ count: 3, next: 8 }),
-      'meta is correct'
-    );
-    assert.strictEqual(responseDocument2.lid, 'https://api.example.com/v1/users', 'lid is correct');
-
-    const structuredDocument2 = store.cache.peekRequest({ lid: 'https://api.example.com/v1/users' });
-    assert.deepEqual(
-      structuredDocument2,
-      {
-        request: { url: 'https://api.example.com/v1/users' },
-        content: {
-          lid: 'https://api.example.com/v1/users',
-          meta: { count: 3, next: 8 },
-        },
-      },
-      'We got the cached structured document back'
-    );
-    const cachedResponse2 = store.cache.peek({ lid: 'https://api.example.com/v1/users' });
-    assert.deepEqual(
-      cachedResponse2,
-      {
-        lid: 'https://api.example.com/v1/users',
-        meta: { count: 3, next: 8 },
-      },
-      'We got the cached response document back'
-    );
-  });
-
-  test('meta documents are correctly cached', function (assert) {
-    const store = this.owner.lookup('service:store') as Store;
-
-    const responseDocument = store.cache.put({
-      request: { url: 'https://api.example.com/v1/users' },
-      content: {
-        meta: { count: 4 },
-      },
-    } as StructuredDocument<ResourceMetaDocument>) as ResourceMetaDocument;
-
-    assert.false('data' in responseDocument, 'No data is associated');
-    assert.deepEqual(responseDocument.meta, { count: 4 }, 'meta is correct');
-    assert.strictEqual(JSON.stringify(responseDocument.meta), JSON.stringify({ count: 4 }), 'meta is correct');
-    assert.strictEqual(responseDocument.lid, 'https://api.example.com/v1/users', 'lid is correct');
-
-    const structuredDocument = store.cache.peekRequest({ lid: 'https://api.example.com/v1/users' });
-    assert.deepEqual(
-      structuredDocument,
-      {
-        request: { url: 'https://api.example.com/v1/users' },
-        content: {
-          lid: 'https://api.example.com/v1/users',
-          meta: { count: 4 },
-        },
-      },
-      'We got the cached structured document back'
-    );
-    const cachedResponse = store.cache.peek({ lid: 'https://api.example.com/v1/users' });
-    assert.deepEqual(
-      cachedResponse,
-      {
-        lid: 'https://api.example.com/v1/users',
-        meta: { count: 4 },
       },
       'We got the cached response document back'
     );
